// frontend/src/app/features/models/usdainfo.model.ts
export interface UsdaInfo {
    data: Datum[];
}

export interface Datum {
    prodn_practice_desc: string;
    domain_desc: string;
    county_name: string;
    freq_desc: string;
    begin_code: string;
    watershed_code: string;
    end_code: string;
    state_alpha: string;
    agg_level_desc: string;
    cV: string;
    state_ansi: string;
    util_practice_desc: string;
    region_desc: string;
    state_fips_code: string;
    county_code: string;
    week_ending: string;
    year: number;
    watershed_desc: string;
    unit_desc: string;
    country_name: string;
    domaincat_desc: string;
    location_desc: string;
    zip_5: string;
    group_desc: string;
    load_time: string;
    value: string;
    asd_desc: string;
    county_ansi: string;
    asd_code: string;
    commodity_desc: string;
    statisticcat_desc: string;
    congr_district_code: string;
    state_name: string;
    reference_period_desc: string;
    source_desc: string;
    class_desc: string;
    sector_desc: string;
    country_code: string;
    short_desc: string;
}